experimance/services/image_server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-server"
version = "0.1.0"
description = "Image server service for the Experimance interactive art installation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ryan Kelln", email = "ryan.kelln@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "experimance-common>=0.1.0",
    "numpy",
    "pyzmq",
    "pillow",
    "opencv-python",
    "aiohttp",
    "requests",
    "fal-client",
    "openai",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "mypy",
    "ruff",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["image_server"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
